Rename connections module to peers
